feat(super search): added base config and component in the draft status
