+++ components/generic/super-search/index.ts
@@ -1,0 +1,1 @@
+export { default as SuperSearch } from './components/super-search'
