+++ components/generic/super-search/types/index.tsx
@@ -1,0 +1,4 @@
+import { GroupEntityType } from '@lib/types/entities/group'
+import { ContentTypeDefinition } from '@lib/types'
+
+export type Data = { groups: GroupEntityType[]; content: ContentTypeDefinition[] }
