+++ lib/api/api-endpoints.js
@@ -26,4 +26,5 @@ export default {
   groups,
   comments: '/api/comments',
   activity: '/api/activity',
+  superSearch: '/api/super-search',
 }
