+++ lib/client/hooks/index.ts
@@ -12,3 +12,4 @@ export * from './use-group-types'
 export { default as useInfinityList } from './use-infinity-list'
 export { default as useFirstMountState } from './use-first-mount-state'
 export { default as useUpdateEffect } from './use-update-effect'
+export { default as useDebounce } from './use-debounce'
