+++ lib/config/load-config.js
@@ -128,6 +128,9 @@ export default function load() {
       'activity.update': newConfig.activity.permissions.update,
       'activity.delete': newConfig.activity.permissions.delete,
       'activity.admin': newConfig.activity.permissions.admin,
+
+      // Super search permissions
+      'superSearch.read': newConfig.superSearch.permissions.read,
     }
 
     // Content permissions
